wip training infrastructure

+++ v2/src/core/analysis/lattice_config.h
@@ -32,7 +32,7 @@ struct alignas(alignof(u64)) ConnectionPtr {
   const ConnectionPtr* previous;
 };
 
-struct alignas(alignof(u64)) LatticeNodePtr {
+struct alignas(alignof(u32)) LatticeNodePtr {
   u16 boundary;
   u16 position;
 };
